Finish tests and polish API.

+++ client_server_admin.go
@@ -58,25 +58,25 @@ type ServerStatistics struct {
 // SystemStats contains statistical data about the system, this is part of
 // ServerStatistics.
 type SystemStats struct {
-	MinorPageFaults     int     `json:"minorPageFaults"`
-	MajorPageFaults     int     `json:"majorPageFaults"`
+	MinorPageFaults     int64   `json:"minorPageFaults"`
+	MajorPageFaults     int64   `json:"majorPageFaults"`
 	UserTime            float64 `json:"userTime"`
 	SystemTime          float64 `json:"systemTime"`
-	NumberOfThreads     int     `json:"numberOfThreads"`
-	ResidentSize        int     `json:"residentSize"`
+	NumberOfThreads     int64   `json:"numberOfThreads"`
+	ResidentSize        int64   `json:"residentSize"`
 	ResidentSizePercent float64 `json:"residentSizePercent"`
-	VirtualSize         int     `json:"virtualSize"`
+	VirtualSize         int64   `json:"virtualSize"`
 }
 
 // Stats is used for various time-related statistics.
 type Stats struct {
-	Sum    int   `json:"sum"`
-	Count  int   `json:"count"`
-	Counts []int `json:"counts"`
+	Sum    float64 `json:"sum"`
+	Count  int64   `json:"count"`
+	Counts []int64 `json:"counts"`
 }
 
 type ClientStats struct {
-	HTTPConnections int   `json:"httpConnections"`
+	HTTPConnections int64 `json:"httpConnections"`
 	ConnectionTime  Stats `json:"connectionTime"`
 	TotalTime       Stats `json:"totalTime"`
 	RequestTime     Stats `json:"requestTime"`
@@ -88,52 +88,52 @@ type ClientStats struct {
 
 // HTTPStats contains statistics about the HTTP traffic.
 type HTTPStats struct {
-	RequestsTotal   int `json:"requestsTotal"`
-	RequestsAsync   int `json:"requestsAsync"`
-	RequestsGet     int `json:"requestsGet"`
-	RequestsHead    int `json:"requestsHead"`
-	RequestsPost    int `json:"requestsPost"`
-	RequestsPut     int `json:"requestsPut"`
-	RequestsPatch   int `json:"requestsPatch"`
-	RequestsDelete  int `json:"requestsDelete"`
-	RequestsOptions int `json:"requestsOptions"`
-	RequestsOther   int `json:"requestsOther"`
+	RequestsTotal   int64 `json:"requestsTotal"`
+	RequestsAsync   int64 `json:"requestsAsync"`
+	RequestsGet     int64 `json:"requestsGet"`
+	RequestsHead    int64 `json:"requestsHead"`
+	RequestsPost    int64 `json:"requestsPost"`
+	RequestsPut     int64 `json:"requestsPut"`
+	RequestsPatch   int64 `json:"requestsPatch"`
+	RequestsDelete  int64 `json:"requestsDelete"`
+	RequestsOptions int64 `json:"requestsOptions"`
+	RequestsOther   int64 `json:"requestsOther"`
 }
 
 // TransactionStats contains statistics about transactions.
 type TransactionStats struct {
-	Started             int `json:"started"`
-	Aborted             int `json:"aborted"`
-	Committed           int `json:"committed"`
-	IntermediateCommits int `json:"intermediateCommits"`
+	Started             int64 `json:"started"`
+	Aborted             int64 `json:"aborted"`
+	Committed           int64 `json:"committed"`
+	IntermediateCommits int64 `json:"intermediateCommits"`
 }
 
 // MemoryStats contains statistics about memory usage.
 type MemoryStats struct {
-	ContextID    int     `json:"contextId"`
+	ContextID    int64   `json:"contextId"`
 	TMax         float64 `json:"tMax"`
-	CountOfTimes int     `json:"countOfTimes"`
-	HeapMax      int     `json:"heapMax"`
-	HeapMin      int     `json:"heapMin"`
+	CountOfTimes int64   `json:"countOfTimes"`
+	HeapMax      int64   `json:"heapMax"`
+	HeapMin      int64   `json:"heapMin"`
 }
 
 // V8ContextStats contains statistics about V8 contexts.
 type V8ContextStats struct {
-	Available int           `json:"available"`
-	Busy      int           `json:"busy"`
-	Dirty     int           `json:"dirty"`
-	Free      int           `json:"free"`
-	Max       int           `json:"max"`
+	Available int64         `json:"available"`
+	Busy      int64         `json:"busy"`
+	Dirty     int64         `json:"dirty"`
+	Free      int64         `json:"free"`
+	Max       int64         `json:"max"`
 	Memory    []MemoryStats `json:"memory"`
 }
 
 // ThreadsStats contains statistics about threads.
 type ThreadStats struct {
-	SchedulerThreads int `json:"scheduler-threads"`
-	Blocked          int `json:"blocked"`
-	Queued           int `json:"queued"`
-	InProgress       int `json:"in-progress"`
-	DirectExec       int `json:"direct-exec"`
+	SchedulerThreads int64 `json:"scheduler-threads"`
+	Blocked          int64 `json:"blocked"`
+	Queued           int64 `json:"queued"`
+	InProgress       int64 `json:"in-progress"`
+	DirectExec       int64 `json:"direct-exec"`
 }
 
 // ServerStats contains statistics about the server.
